Conflicts:
	Code/EnvironmentalApp/EnvironmentalApp.Core.Tests/obj/Debug/EnvironmentalApp.Core.Tests.dll
	Code/EnvironmentalApp/EnvironmentalApp.Core/obj/Debug/EnvironmentalApp.Core.csproj.FileListAbsolute.txt
	Code/EnvironmentalApp/EnvironmentalApp.Core/obj/Debug/EnvironmentalApp.Core.dll
	Code/EnvironmentalApp/EnvironmentalApp.Data.Tests/obj/Debug/EnvironmentalApp.Data.Tests.csproj.FileListAbsolute.txt
	Code/EnvironmentalApp/EnvironmentalApp.Data.Tests/obj/Debug/EnvironmentalApp.Data.Tests.dll
	Code/EnvironmentalApp/EnvironmentalApp.Data/obj/Debug/EnvironmentalApp.Data.csproj.FileListAbsolute.txt
	Code/EnvironmentalApp/EnvironmentalApp.Data/obj/Debug/EnvironmentalApp.Data.dll
	Code/EnvironmentalApp/EnvironmentalApp.Gui/Controllers/HomeController.cs
	Code/EnvironmentalApp/EnvironmentalApp.Gui/obj/Debug/EnvironmentalApp.Gui.csproj.FileListAbsolute.txt
	Code/EnvironmentalApp/EnvironmentalApp.Gui/obj/Debug/EnvironmentalApp.Gui.dll
	Code/EnvironmentalApp/EnvironmentalApp.Services/obj/Debug/EnvironmentalApp.Services.csproj.FileListAbsolute.txt
	Code/EnvironmentalApp/EnvironmentalApp.Services/obj/Debug/EnvironmentalApp.Services.dll

+++ Code/EnvironmentalApp/EnvironmentalApp.Core/Models/BaseModel.cs
@@ -11,7 +11,7 @@ namespace EnvironmentalApp.Core.Models
         public Guid Id { get; set; }
         public DateTime TimeStamp { get; set; }
         public DateTime ReadingDateTime { get; set; }
-        public decimal Reading { get; set; }
+        public float Reading { get; set; }
         public int TimeStep { get; set; }
         public int Status { get; set; }
